+++ Magic/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 7.8.4
 
+ - Use SkillAPI allies for friendly fire checks
  - Fix "requires" items in Citizens traits getting removed on restart
  - Add "language" config option, to easily change the language of in-game text.
    Thanks to the generous help of translators, the following built-in languages are available: